reguisition-system

+++ app/Http/Livewire/RequestMaterialComponent.php
@@ -3,11 +3,14 @@
 namespace App\Http\Livewire;
 
 use App\Models\MaterialRequest;
+use App\Models\MaterialRequestStatus;
+use Illuminate\Support\Facades\Auth;
 use Livewire\Component;
 
 class RequestMaterialComponent extends Component
 {
-    public $material;
+    protected $listeners = ['clear', 'setStatus'];
+    public $material, $comment, $statusId;
 
     public function rules()
     {
@@ -16,7 +19,18 @@ class RequestMaterialComponent extends Component
             'material.description' => 'required|string',
             'material.qty' => 'required|numeric',
             'material.type' => 'required|string',
+            'comment' => 'nullable|string|max:255'
         ];
+    }
+
+    public function updated($property, $value)
+    {
+        $this->validateOnly($property);
+    }
+
+    public function clear()
+    {
+        $this->reset('comment');
     }
 
     public function store()
@@ -28,27 +42,48 @@ class RequestMaterialComponent extends Component
             : $this->emit('toast', 'error', "Failed to submit your request, please try again.", "Material Request");
     }
 
+    public function openModal($id)
+    {
+        $this->comment = MaterialRequestStatus::where('updated_by', auth()->id())->where('material_request_id', $id)->value('comments');
+        $this->emit('showModal', 'update_comments');
+    }
+
     public function setStatus($id, $value)
     {
-        MaterialRequest::find($id)->update([
-            'status' => $value
-        ])
+        $status = auth()->user()->materialRequestStatuses()->updateOrCreate([
+            'material_request_id' => $id,
+            'designation' => auth()->user()->getDesignation(),
+        ], [
+            'status' => $value == 'Approved' ? true : false,
+        ]);
+        $status
             ? $this->emit('toast', 'success', "Request status was updated successfully. 😉", "Material Request Status")
             : $this->emit('toast', 'error', "Failed to update status 😉", "Material Request Status");
+
+        $this->statusId = $status->id;
+        $this->openModal($id);
     }
+
+    public function saveRemarks()
+    {
+        MaterialRequestStatus::find($this->statusId)->update([
+            'comments' => $this->comment
+        ]);
+
+        $this->emit('closeModal', 'update_comments');
+    }
+
     public function delete($id)
     {
-        $model = MaterialRequest::find($id);
-        if (!is_null($model)) {
-            $model->delete()
+        if ($model = MaterialRequest::find($id))
+            $model->statuses()->delete() && $model->delete()
                 ? $this->emit('toast', 'success', "Request was deleted successfully. 😉", "Material Request")
                 : $this->emit('toast', 'error', "Failed to delete request 😉", "Material Request");
-        }
     }
 
     public function render()
     {
-        auth()->user()->role == 'admin' || auth()->user()->role == 'manager'
+        in_array(auth()->user()->role, ['admin', 'manager', 'ceo', 'finance-admin', 'finance-dept'])
             ? $requests = MaterialRequest::paginate(20)
             : $requests = MaterialRequest::where('employee_id', auth()->user()->employee->id)->paginate(20);
